+++ CHANGELOG.md
@@ -14,11 +14,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Transferable objects in call arguments and return values are automatically added to `TransferList`, by [@compulim](https://github.com/compulim) in PR [#XXX](https://github.com/compulim/message-port-rpc/pull/XXX)
+- Transferable objects in call arguments and return values are automatically added to `TransferList`, by [@compulim](https://github.com/compulim) in PR [#65](https://github.com/compulim/message-port-rpc/pull/65)
 
 ### Removed
 
-- 💥 `transfer` in call `Options` is removed in favor of the auto-transfer feature, by [@compulim](https://github.com/compulim) in PR [#XXX](https://github.com/compulim/message-port-rpc/pull/XXX)
+- 💥 `transfer` in call `Options` is removed in favor of the auto-transfer feature, by [@compulim](https://github.com/compulim) in PR [#65](https://github.com/compulim/message-port-rpc/pull/65)
 
 ### Removed