+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 | 👷🏻   | Related to development experience and non-production impacting |
 | 💥   | Related to API breaking changes                                |
 
-## [Unreleased]
+## [3.0.0] - 2026-06-11
 
 ### Added
 
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - 💥 `transfer` in call `Options` is removed in favor of the auto-transfer feature, by [@compulim](https://github.com/compulim) in PR [#65](https://github.com/compulim/message-port-rpc/pull/65)
+
 ## [2.0.1] - 2026-05-31
 
 ### Changed
@@ -187,7 +188,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First public release
 
-[Unreleased]: https://github.com/compulim/message-port-rpc/compare/v2.0.1...HEAD
+[3.0.0]: https://github.com/compulim/message-port-rpc/compare/v2.0.1...v3.0.0
 [2.0.1]: https://github.com/compulim/message-port-rpc/compare/v2.0.0...v2.0.1
 [2.0.0]: https://github.com/compulim/message-port-rpc/compare/v1.0.1...v2.0.0
 [1.0.1]: https://github.com/compulim/message-port-rpc/compare/v1.0.0...v1.0.1